move billing under ee
