+++ packages/web/src/ee/features/billing/components/changeBillingEmailCard.tsx
@@ -1,11 +1,11 @@
 "use client"
 
-import { changeSubscriptionBillingEmail } from "@/actions"
 import { useToast } from "@/components/hooks/use-toast"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { Form, FormControl, FormField, FormItem, FormMessage } from "@/components/ui/form"
 import { Input } from "@/components/ui/input"
+import { changeSubscriptionBillingEmail } from "@/ee/features/billing/actions"
 import useCaptureEvent from "@/hooks/useCaptureEvent"
 import { useDomain } from "@/hooks/useDomain"
 import { isServiceError } from "@/lib/utils"
