+++ packages/web/src/ee/features/billing/stripe.ts
@@ -1,8 +1,9 @@
 import 'server-only';
 import { env } from '@/env.mjs'
 import Stripe from "stripe";
+import { hasEntitlement } from '@/features/entitlements/server';
 
-export const IS_BILLING_ENABLED = env.STRIPE_SECRET_KEY !== undefined;
+export const IS_BILLING_ENABLED = hasEntitlement('billing') && env.STRIPE_SECRET_KEY !== undefined;
 
 export const stripeClient =
   IS_BILLING_ENABLED
